+++ packages/core-ui/command-surface/src/components/modals/device-control-graph/ControlGraph.tsx
@@ -5,14 +5,18 @@ import React, { useEffect, useMemo, useState } from "react";
 import { unit as mathUnit } from 'mathjs';
 
 export const ControlGraphModal = (props: {tags: HMITag[], types: HMIType[], selected?: any, open: boolean, onClose?: () => void, onSubmit?: (graph: any) => void}) => {
+  
   const [graph, setGraph] = useState<{
     unit?: string;
     deviceID?: string;
     keyID?: string;
     totalize?: boolean;
     timeBucket?: string;
+    xAxisDomain?: [string, string],
+    yAxisDomain?: [string, string],
   }>({});
 
+  const customAxis = graph.xAxisDomain != undefined && graph.yAxisDomain != undefined
   
   useEffect(() => {
     setGraph({
@@ -114,6 +118,54 @@ export const ControlGraphModal = (props: {tags: HMITag[], types: HMIType[], sele
             onChange={(e) => setGraph({ ...graph, timeBucket: e.target.value })} />
      
           <FormControlLabel label="Totalise" control={<Checkbox checked={graph.totalize || false} onChange={(e) => setGraph({...graph, totalize: e.target.checked})} />} />
+
+          <Divider />
+          
+          <FormControlLabel label="Custom Axis" control={<Checkbox checked={customAxis} onChange={(e, checked) => {
+            if(checked){
+               setGraph({...graph, xAxisDomain: ['0', 'auto'], yAxisDomain: ['0', 'auto']})
+            }else{
+              setGraph({...graph, xAxisDomain: undefined, yAxisDomain: undefined})
+            }
+          }}/>} />
+          
+          {customAxis ? (
+            <Box sx={{display: 'flex', flexDirection: "column"}}>
+              <Divider />
+              <Box sx={{display: 'flex', marginBottom: '12px'}}>
+                <TextField 
+                  sx={{marginRight: '6px'}}
+                  fullWidth 
+                  size="small" 
+                  onChange={(e) => setGraph({...graph, xAxisDomain: [ e.target.value, graph.xAxisDomain?.[1] || ''] })} 
+                  value={graph.xAxisDomain?.[0] || ''} 
+                  label={"X Start"} /> 
+                <TextField 
+                  sx={{marginLeft: '6px'}}
+                  fullWidth 
+                  size="small" 
+                  onChange={(e) => setGraph({...graph, xAxisDomain: [ graph.xAxisDomain?.[0] || '',  e.target.value ] })} 
+                  value={graph.xAxisDomain?.[1] || ''} 
+                  label={"X End"} />
+              </Box>
+              <Box sx={{display: 'flex'}}>
+                <TextField 
+                  sx={{marginRight: '6px'}}
+                  fullWidth 
+                  size="small" 
+                  onChange={(e) => setGraph({...graph, yAxisDomain: [ e.target.value, graph.yAxisDomain?.[1] || ''] })} 
+                  value={graph.yAxisDomain?.[0] || ''} 
+                  label={"Y Start"} /> 
+                <TextField
+                  sx={{marginLeft: '6px'}}
+                  fullWidth
+                  size="small" 
+                  onChange={(e) => setGraph({...graph, yAxisDomain: [ graph.yAxisDomain?.[0] || '',  e.target.value ] })} 
+                  value={graph.yAxisDomain?.[1] || ''} 
+                  label={"Y End"} />
+              </Box>
+            </Box>
+          ) : null}
         </Box>
       </DialogContent>
       <DialogActions>