+++ packages/communication/opc-server/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.313](https://github.com/TheTechCompany/sudbuster/compare/v1.4.15-alpha.312...v1.4.15-alpha.313) (2024-09-17)
+
+**Note:** Version bump only for package @hive-command/opcua-server
+
+
+
+
+
+## [1.4.15-alpha.312](https://github.com/TheTechCompany/sudbuster/compare/v1.4.15-alpha.311...v1.4.15-alpha.312) (2024-09-11)
+
+**Note:** Version bump only for package @hive-command/opcua-server
+
+
+
+
+
+## [1.4.15-alpha.311](https://github.com/TheTechCompany/sudbuster/compare/v1.4.15-alpha.310...v1.4.15-alpha.311) (2024-05-23)
+
+**Note:** Version bump only for package @hive-command/opcua-server
+
+
+
+
+
+## [1.4.15-alpha.310](https://github.com/TheTechCompany/sudbuster/compare/v1.4.15-alpha.309...v1.4.15-alpha.310) (2024-05-23)
+
+**Note:** Version bump only for package @hive-command/opcua-server
+
+
+
+
+
 ## [1.4.15-alpha.309](https://github.com/TheTechCompany/sudbuster/compare/v1.4.15-alpha.308...v1.4.15-alpha.309) (2024-04-29)
 
 **Note:** Version bump only for package @hive-command/opcua-server